feat: View browsing history
Add a new history page which allows users to view their browsing history.
This new page is backed by a new managed api `/api/history` which in turn consumes the Browsing history API (served by the history-api Function App).

Instead of using real authentication (e.g. using the integrated auth features of Static Web Apps), we opted instead for a login simulation (based on the browser's session storage) in order to reduce the complexity of testing the entire setup for the HoL lab on the D-Day and prevent issues related to not having the required permissions to use AAD.

Fixes #22